+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\CartController;
 use App\Http\Controllers\ColorController;
+use App\Http\Controllers\OrderController;
 use App\Http\Controllers\PaymentStatusController;
 use App\Http\Controllers\ProductController;
 use App\Http\Controllers\CategoryController;
@@ -16,46 +17,49 @@ use App\Http\Middleware\CheckRole;
 
 Route::get('/', [ProductController::class, 'showProducts'])->name('welcome');
 
-Route::get('/dashboard', function () {
+Route::get('/admin/dashboard', function () {
     return Inertia::render('Dashboard');
 })->middleware(['auth', 'verified', CheckRole::class . ':admin'])->name('dashboard');
 
 Route::middleware(['auth', 'verified', CheckRole::class . ':admin'])->group(function () {
-    Route::get('/products', [ProductController::class, 'index'])->name('products.index');
-    Route::get('/products/create', [ProductController::class, 'create'])->name('products.create');
-    Route::post('/products', [ProductController::class, 'store'])->name('products.store');
-    Route::get('/products/{product}/edit', [ProductController::class, 'edit'])->name('products.edit');
-    Route::post('/products/{product}', [ProductController::class, 'update'])->name('products.update');
-    Route::get('/products/{product}/delete', [ProductController::class, 'delete'])->name('products.delete');
-    Route::delete('/products/{product}', [ProductController::class, 'destroy'])->name('products.destroy');
+    Route::get('/admin/products', [ProductController::class, 'index'])->name('products.index');
+    Route::get('/admin/products/create', [ProductController::class, 'create'])->name('products.create');
+    Route::post('/admin/products', [ProductController::class, 'store'])->name('products.store');
+    Route::get('/admin/products/{product}/edit', [ProductController::class, 'edit'])->name('products.edit');
+    Route::post('/admin/products/{product}', [ProductController::class, 'update'])->name('products.update');
+    Route::get('/admin/products/{product}/delete', [ProductController::class, 'delete'])->name('products.delete');
+    Route::delete('/admin/products/{product}', [ProductController::class, 'destroy'])->name('products.destroy');
 });
 
 Route::get('/products/{product}', [ProductController::class, 'show'])->name('products.show');
 
 Route::middleware(['auth', 'verified', CheckRole::class . ':admin'])->group(function () {
-    Route::get('/categories', [CategoryController::class, 'index'])->name('categories.index');
-    Route::get('/categories/create', [CategoryController::class, 'create'])->name('categories.create');
-    Route::post('/categories', [CategoryController::class, 'store'])->name('categories.store');
-    Route::get('/categories/{category}/edit', [CategoryController::class, 'edit'])->name('categories.edit');
-    Route::put('/categories/{category}', [CategoryController::class, 'update'])->name('categories.update');
-    Route::delete('/categories/{category}', [CategoryController::class, 'destroy'])->name('categories.destroy');
-    Route::get('/categories/{category}/delete', [CategoryController::class, 'delete'])->name('categories.delete');
+    Route::get('/admin/categories', [CategoryController::class, 'index'])->name('categories.index');
+    Route::get('/admin/categories/create', [CategoryController::class, 'create'])->name('categories.create');
+    Route::post('/admin/categories', [CategoryController::class, 'store'])->name('categories.store');
+    Route::get('/admin/categories/{category}/edit', [CategoryController::class, 'edit'])->name('categories.edit');
+    Route::put('/admin/categories/{category}', [CategoryController::class, 'update'])->name('categories.update');
+    Route::delete('/admin/categories/{category}', [CategoryController::class, 'destroy'])->name('categories.destroy');
+    Route::get('/admin/categories/{category}/delete', [CategoryController::class, 'delete'])->name('categories.delete');
 
-    Route::get('/colors', [ColorController::class, 'index'])->name('colors.index');
-    Route::get('/colors/create', [ColorController::class, 'create'])->name('colors.create');
-    Route::post('/colors', [ColorController::class, 'store'])->name('colors.store');
-    Route::get('/colors/{color}/edit', [ColorController::class, 'edit'])->name('colors.edit');
-    Route::put('/colors/{color}', [ColorController::class, 'update'])->name('colors.update');
-    Route::delete('/colors/{color}', [ColorController::class, 'destroy'])->name('colors.destroy');
-    Route::get('/colors/{color}/delete', [ColorController::class, 'delete'])->name('colors.delete');
+    Route::get('/admin/colors', [ColorController::class, 'index'])->name('colors.index');
+    Route::get('/admin/colors/create', [ColorController::class, 'create'])->name('colors.create');
+    Route::post('/admin/colors', [ColorController::class, 'store'])->name('colors.store');
+    Route::get('/admin/colors/{color}/edit', [ColorController::class, 'edit'])->name('colors.edit');
+    Route::put('/admin/colors/{color}', [ColorController::class, 'update'])->name('colors.update');
+    Route::delete('/admin/colors/{color}', [ColorController::class, 'destroy'])->name('colors.destroy');
+    Route::get('/admin/colors/{color}/delete', [ColorController::class, 'delete'])->name('colors.delete');
 
-    Route::get('/sizes', [SizeController::class, 'index'])->name('sizes.index');
-    Route::get('/sizes/create', [SizeController::class, 'create'])->name('sizes.create');
-    Route::post('/sizes', [SizeController::class, 'store'])->name('sizes.store');
-    Route::get('/sizes/{size}/edit', [SizeController::class, 'edit'])->name('sizes.edit');
-    Route::put('/sizes/{size}', [SizeController::class, 'update'])->name('sizes.update');
-    Route::delete('/sizes/{size}', [SizeController::class, 'destroy'])->name('sizes.destroy');
-    Route::get('/sizes/{size}/delete', [SizeController::class, 'delete'])->name('sizes.delete');
+    Route::get('/admin/sizes', [SizeController::class, 'index'])->name('sizes.index');
+    Route::get('/admin/sizes/create', [SizeController::class, 'create'])->name('sizes.create');
+    Route::post('/admin/sizes', [SizeController::class, 'store'])->name('sizes.store');
+    Route::get('/admin/sizes/{size}/edit', [SizeController::class, 'edit'])->name('sizes.edit');
+    Route::put('/admin/sizes/{size}', [SizeController::class, 'update'])->name('sizes.update');
+    Route::delete('/admin/sizes/{size}', [SizeController::class, 'destroy'])->name('sizes.destroy');
+    Route::get('/admin/sizes/{size}/delete', [SizeController::class, 'delete'])->name('sizes.delete');
+
+    Route::get('/admin/orders', [OrderController::class, 'index'])->name('admin.orders.index');
+    Route::get('/admin/orders/{order}', [OrderController::class, 'show'])->name('orders.show');
 });
 
 Route::middleware('auth')->group(function () {
@@ -75,7 +79,7 @@ Route::middleware('auth')->group(function () {
 // Rutas para el proceso de checkout
 Route::middleware('auth')->group(function () {
     Route::get('/checkout', [CheckoutController::class, 'index'])->name('checkout.index'); 
-    Route::get('/checkout/payment', [CheckoutController::class, 'payment'])->name('checkout.payment'); 
+    Route::post('/checkout/payment', [CheckoutController::class, 'payment'])->name('checkout.payment'); 
 
     Route::get('/payment/success', [PaymentStatusController::class, 'success'])->name('payment.success');
     Route::get('/payment/failure', [PaymentStatusController::class, 'failure'])->name('payment.failure');
